Bugfix: Users cannot delete other users posts
Delete button removed from posts on home page.

Identity verification added to delete post function.

+++ functions/delete_post.php
@@ -4,18 +4,37 @@ include("includes/connection.php");
 
 if (isset($_GET['post_id'])) {
 
+
     $post_id = $_GET['post_id'];
+    $get_post_owner = "select owner_id from posts where post_id = '$post_id'";
+    $post_owner_id = mysqli_query($con, $get_post_owner);
 
-    $delete_post = "delete from posts where post_id='$post_id'";
-    $run_delete = mysqli_query($con, $delete_post);
+    if (is_owner($post_owner_id)) {
+        $delete_post = "delete from posts where post_id='$post_id'";
+        $run_delete = mysqli_query($con, $delete_post);
 
-    if ($run_delete) {
-        echo "<script>alert('This post has been deleted!')</script>";
-        echo "<script>window.open('../user_posts.php','_self')</script>";
+        if ($run_delete) {
+            echo "<script>alert('This post has been deleted!')</script>";
+            echo "<script>window.open('../user_posts.php','_self')</script>";
+        }
+        else {
+            echo "<script>alert('Could not contact server')</script>";
+        }
     }
     else {
-        echo "<script>alert('Could not contact server')</script>";
+        echo "<script>alert('You cannot delete another users post!')</script>";
     }
+}
+
+function is_owner($post_owner_id)
+{
+    $user = $_SESSION['user_email'];
+    $get_user = "select * from users where user_email = '$user'";
+    $run_user = mysqli_query($con, $get_user);
+    $row = mysqli_fetch_array($run_user);
+    $user_id = $row['user_id'];
+
+    return $user_id == $post_owner_id ? true : false;
 }
 
 ?>
